requirements.txt
fastapi 
uvicorn 
langchain 
langchain-community 
python-pptx 
pydantic
langgraph 